# headerfs: fail gracefully on header write