--- a/headerfs/store.go
+++ b/headerfs/store.go
@@ -3,6 +3,7 @@ package headerfs
 import (
 	"bytes"
 	"fmt"
+	"io"
 	"os"
 	"path/filepath"
 	"sync"
@@ -84,6 +85,23 @@ var headerBufPool = sync.Pool{
 	New: func() interface{} { return new(bytes.Buffer) },
 }
 
+// FileInterface defines the minimum file operations needed by headerStore.
+type FileInterface interface {
+	// Basic I/O operations.
+	io.Reader
+	io.Writer
+	io.Closer
+
+	// Extended I/O positioning.
+	io.Seeker
+	io.ReaderAt
+
+	// File-specific operations.
+	Stat() (os.FileInfo, error)
+	Sync() error
+	Truncate(size int64) error
+}
+
 // headerStore combines a on-disk set of headers within a flat file in addition
 // to a database which indexes that flat file. Together, these two abstractions
 // can be used in order to build an indexed header store for any type of
@@ -96,7 +114,7 @@ type headerStore struct {
 
 	fileName string
 
-	file *os.File
+	file FileInterface
 
 	*headerIndex
 }