## Description

This PR improves error handling by failing gracefully when header writes fail. This helps prevent unexpected `EOF` errors during the process.