@mohamedawnallah 2025-05-13
> Can we make a basic benchmark here to gauge if this affects write perf in a traditional set up?

That makes sense, thanks. I have added that `BenchmarkHeaderStoreAppendRaw` benchmarking test. Here is the report:

## Without this PR (commit afcfeb1)
$ go test -bench=BenchmarkHeaderStoreAppendRaw -benchmem
goos: linux
goarch: amd64
pkg: github.com/lightninglabs/neutrino/headerfs
cpu: AMD EPYC-Rome-v4 Processor (no XSAVES)
BenchmarkHeaderStoreAppendRaw-4          1083670              1214 ns/op               0 B/op          0 allocs/op
PASS
ok      github.com/lightninglabs/neutrino/headerfs      2.521s


## With this PR
$ go test -bench=BenchmarkHeaderStoreAppendRaw -benchmem
goos: linux
goarch: amd64
pkg: github.com/lightninglabs/neutrino/headerfs
cpu: AMD EPYC-Rome-v4 Processor (no XSAVES)
BenchmarkHeaderStoreAppendRaw-4           951910              1305 ns/op               0 B/op          0 allocs/op
PASS
ok      github.com/lightninglabs/neutrino/headerfs      2.401s


## Benchmarking Report - Generated by Gemini

Based on the benchmark results:

**Before the change:**
- 1,083,670 operations
- 1,214 ns/op
- 0 B/op (no memory allocations)

**After the change:**
- 951,910 operations
- 1,305 ns/op
- 0 B/op (no memory allocations)

**Analysis:**
1. **Performance impact:** The new implementation is about 7.5% slower (1,305 ns vs 1,214 ns per operation)
2. **Throughput decrease:** The number of operations completed decreased by about 12.2%
3. **Memory usage:** Both implementations are equally memory-efficient with zero allocations

**Conclusion:**
The added error handling with the `Seek` operation does introduce a small performance overhead of approximately 91 nanoseconds per operation. Given that:

1. This is a very modest performance impact (~7.5%)
2. The code now properly handles partial writes and maintains data integrity
3. The error messages are much more informative
4. There's still no memory allocation overhead

The tradeoff appears reasonable. The performance cost is minimal while the reliability improvements are significant. For a storage operation where data integrity is critical, this seems like an acceptable performance trade-off to gain much better error handling and recovery capability.