@mohamedawnallah 2025-05-20
From the benchmarking results, it appears there's overhead on the seek operation. However, this is somehow misleading because the benchmark calls `appendRaw` on only a single header (creating a one-to-one mapping between seek operations and headers). In the actual codebase, this operation is performed on batches, meaning one seek operation handles multiple headers. Therefore, the observed 7.5% performance impact is significantly lower in real-world usage.

I think integrating this code change in the codebase would be fine. What do you think @Roasbeef?