@ziggie1984 2025-06-02
Added the column linter here: https://github.com/lightninglabs/neutrino/pull/318